pyproject.toml
[build-system]
requires      = ["setuptools>=69", "wheel>=0.42"]
build-backend = "setuptools.build_meta"

[project]
name = "justai"
version = "3.0.-1"
description = "Makes working with OpenAI's GPT API and other LLM's super easy"
readme = "README.md"
authors = [{ name = "HP Harmsen", email = "hp@harmsen.nl" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["ChatGPT", "GPT4", "GPT3.5", "api", "Claude", "Anthropic", "Lllama"]
dependencies = [
    "openai",
    "tenacity",
    "python-dateutil",
    "python-dotenv",
    "rich",
    "requests",
    "tiktoken",
    "lxml",
    "llama-cpp-python",
    "anthropic"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["black", "pytest", "build", "twine", "nox"]

[project.urls]
Homepage = "https://github.com/hpharmsen/justai"
